refactor: WeightChart.tsx 리펙토링 (#84)
