+++ src/features/health/constants/몇째주.ts
@@ -1,0 +1,1 @@
+export const 몇째주 = ['첫째주', '둘째주', '셋째주', '넷째주', '다섯째주']
